+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to the CodinGame extension will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.1] - 2024-05-05
+
+### Changed
+- Updated extension to vscode 1.88 version
+
+### Removed
+- Send bot code and get CodinGame Id commands were removed, as the CodinGame API changes too often to rely on it
+- Removed settings associated with such commands
+
+### Fixed
+- Detect when we are trying to create a bot with an already used bot name.
+
 ## [1.0.2] - 2021-03-06
 
 ### Added